refactor: 유틸 패키지 분리
