+++ BE/side-dish-backend/src/main/java/com/codesquad/team14/utils/Badge.java
@@ -1,4 +1,4 @@
-package com.codesquad.team14.domain;
+package com.codesquad.team14.utils;
 
 public enum Badge {
     EVENT, LAUNCHING, BEST
